Added calculation of Pet age to pass to viewmodel to rend the age in details page.
Integrated in search model a multi selection dropdown list and ability to easily remove the selection ( use on color and race ) keep the rest as checks.

TODO // OnSelect method of datepicked is not firing properly. This needs to be further investigated as this is the way to grey out the hours of timepicker when an appointment has already been confirmed...